Add tests exercising the commit offset asynchronous API with unknown topic ids.
