+++ clients/src/test/java/org/apache/kafka/clients/admin/internals/AlterConsumerGroupOffsetsHandlerTest.java
@@ -17,7 +17,7 @@
 package org.apache.kafka.clients.admin.internals;
 
 import static java.util.Collections.emptyMap;
-import static org.apache.kafka.test.MoreAssertions.assertRequestEquals;
+import static org.apache.kafka.common.requests.OffsetCommitRequestTest.assertRequestEquals;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static java.util.Collections.singleton;
 import static java.util.Collections.singletonList;
